+++ tests/python/test_dataset_aggregation.py
@@ -117,6 +117,14 @@ def test_aggregation(context):
             f"The list length of {list_length} does not match expected "
             f"value of {actual_length}.")
     log_data(context, LLDebug, "List length check")
+    
+    # Check the return of a range of datasets from the aggregated list
+    num_datasets = client.get_dataset_list_range(list_name, 0, 1)
+    if (len(num_datasets) != 2):
+        raise RuntimeError(
+            f"The length is {len(num_datasets)}, which does not "
+            f"match expected value of 2.")
+    log_data(context, LLDebug, "Retrieve datasets from list checked")
 
     # Retrieve datasets via the aggregation list
     datasets = client.get_datasets_from_list(list_name)
@@ -126,7 +134,31 @@ def test_aggregation(context):
             f"does not match expected value of {list_length}.")
     for ds in datasets:
         check_dataset(ds)
-    log_data(context, LLDebug, "DataSet retrieval")
+    log_data(context, LLDebug, "DataSet list retrieval")
+    
+    # Rename a list of datasets
+    client.rename_list(list_name, "new_list_name")
+    renamed_list_datasets = client.get_datasets_from_list("new_list_name")
+    if len(renamed_list_datasets) != list_length:
+        raise RuntimeError(
+            f"The number of datasets received {len(datasets)} "
+            f"does not match expected value of {list_length}.")
+    for ds in renamed_list_datasets:
+        check_dataset(ds)
+    log_data(context, LLDebug, "DataSet list rename complete")
+    
+    # Copy a list of datasets
+    client.copy_list("new_list_name", "copied_list_name")
+    copied_list_datasets = client.get_datasets_from_list("copied_list_name")
+    if len(copied_list_datasets) != list_length:
+        raise RuntimeError(
+            f"The number of datasets received {len(datasets)} "
+            f"does not match expected value of {list_length}.")
+    for ds in copied_list_datasets:
+        check_dataset(ds)
+    log_data(context, LLDebug, "DataSet list copied")
+    
+    
 
 # ------------ helper functions ---------------------------------